requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asteval==1.0.6
astropy==7.0.1
astropy-iers-data==0.2025.4.7.0.35.30
astroquery==0.4.10
beautifulsoup4==4.13.3
bidict==0.23.1
blinker==1.9.0
blosc2==3.3.0
caugetch==0.0.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
clipboard==0.0.4
colorama==0.4.6
configparser==7.2.0
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
dill==0.3.9
Django==5.2
fastapi==0.115.12
Flask==3.1.0
Flask-SocketIO==5.5.1
fonttools==4.57.0
getpass4==0.0.14.1
gunicorn==23.0.0
h11==0.14.0
h5py==3.13.0
habanero==2.2.0
hitran-api==1.2.2.3
hjson==3.1.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.9.0
Jinja2==3.1.6
joblib==1.4.2
json-tricks==3.17.3
keyring==25.6.0
kiwisolver==1.4.8
llvmlite==0.44.0
lmfit==1.3.3
lxml==5.3.2
MarkupSafe==3.0.2
matplotlib==3.10.1
more-itertools==10.6.0
mpldatacursor==0.7.1
msgpack==1.1.0
narwhals==1.34.1
ndindex==1.9.2
numba==0.61.2
numexpr==2.10.2
numpy==2.2.4
packaging==24.2
pandas==2.2.3
PeakUtils==1.3.5
periodictable==2.0.2
pillow==11.1.0
platformdirs==4.3.7
plotly==6.0.1
psutil==7.0.0
publib==0.4.0
py-cpuinfo==9.0.0
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
pyerfa==2.0.1.5
PyMySQL==1.1.1
pyparsing==3.2.3
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.0
pytz==2025.2
pyvo==1.6.2
PyYAML==6.0.2
radis==0.16
redis==5.2.1
requests==2.32.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scipy==1.15.2
seaborn==0.13.2
SecretStorage==3.3.3
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
starlette==0.46.1
tables==3.10.2
termcolor==3.0.1
toml==0.10.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
uncertainties==3.2.2
urllib3==2.2.0
uvicorn==0.34.0
vulkan==1.3.275.1
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
